Remove TOS class and fix the SectionColorsTest.
