--- app/src/main/java/com/github/campus_capture/bootcamp/authentication/TOS.java
@@ -1,8 +1,0 @@
-package com.github.campus_capture.bootcamp.authentication;
-
-public class TOS {
-    public static final String TEXT =
-        "By accepting those terms of service you agree that your data will be collected and used for non-legal activities such as DDoS-ing, Botting and...";
-
-    public static Boolean asAgreed;
-}
